[type:refactor] refactor shenyu logging module
