+++ shenyu-plugin/shenyu-plugin-logging/shenyu-plugin-logging-rocketmq/src/main/java/org/apache/shenyu/plugin/logging/rocketmq/client/RocketMQLogCollectClient.java
@@ -29,13 +29,15 @@ import org.apache.shenyu.plugin.logging.common.client.LogConsumeClient;
 import org.apache.shenyu.plugin.logging.common.constant.GenericLoggingConstant;
 import org.apache.shenyu.plugin.logging.common.entity.LZ4CompressData;
 import org.apache.shenyu.plugin.logging.common.entity.ShenyuRequestLog;
+import org.apache.shenyu.plugin.logging.common.utils.LogCollectConfigUtils;
 import org.apache.shenyu.plugin.logging.rocketmq.config.RocketMQLogCollectConfig;
-import org.apache.shenyu.plugin.logging.rocketmq.utils.RocketLogCollectConfigUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.nio.charset.StandardCharsets;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Objects;
 import java.util.Properties;
 import java.util.concurrent.atomic.AtomicBoolean;
@@ -54,6 +56,8 @@ public class RocketMQLogCollectClient implements LogConsumeClient {
     private String topic;
 
     private final AtomicBoolean isStarted = new AtomicBoolean(false);
+
+    private static Map<String, String> apiTopicMap = new HashMap<>();
 
     /**
      * init producer.
@@ -101,7 +105,7 @@ public class RocketMQLogCollectClient implements LogConsumeClient {
             return;
         }
         logs.forEach(log -> {
-            String logTopic = StringUtils.defaultIfBlank(RocketLogCollectConfigUtils.getTopic(log.getPath()), topic);
+            String logTopic = StringUtils.defaultIfBlank(LogCollectConfigUtils.getTopic(log.getPath(), apiTopicMap), topic);
             try {
                 producer.sendOneway(toMessage(logTopic, log));
             } catch (Exception e) {
@@ -127,6 +131,13 @@ public class RocketMQLogCollectClient implements LogConsumeClient {
         return compressor.compress(srcByte);
     }
 
+    /**
+     * set api topic map.
+     * @param uriTopicMap api topic map
+     */
+    public static void setTopic(final Map<String, String> uriTopicMap) {
+        apiTopicMap = uriTopicMap;
+    }
 
     /**
      * close producer.
